Renderer: remove VertexBufferDesc, implement VertexFormatDesc
- Move VertexBufferDesc methods to VertexFormatDesc.
- No attempt made at present to migrate callsites.
- Allow querying attribute entries in a VertexFormatDesc created from an AttributeSet by VertexAttrib semantic.
- Expand VertexAttrib to 32 bits again, as it is no longer a member of a vertex format descriptor.

+++ src/graphics/Types.h
@@ -9,7 +9,7 @@
 namespace Graphics {
 
 	//Vertex attribute semantic
-	enum VertexAttrib : uint8_t {
+	enum VertexAttrib : uint32_t {
 		ATTRIB_NONE = 0,
 		ATTRIB_POSITION = (1u << 0),
 		ATTRIB_NORMAL = (1u << 1),
